refactor(lint,candidates): split rules.ts and candidates.ts under the size limit (Phase 4 groundwork)
